BUG: WriteToFile() in the resampler base wrote a null character for unsigned char data when DefaultPixelValue = 0
Increased the elastix version to 4.505

+++ src/Core/ComponentBaseClasses/elxResamplerBase.hxx
@@ -450,7 +450,8 @@ ResamplerBase<TElastix>
 
   /** Write the DefaultPixelValue. */
   xl::xout["transpar"] << "(DefaultPixelValue "
-    << this->GetAsITKBaseType()->GetDefaultPixelValue() << ")" << std::endl;
+    << static_cast<double>( this->GetAsITKBaseType()->GetDefaultPixelValue() )
+    << ")" << std::endl;
 
   /** Write the output image format. */
   std::string resultImageFormat = "mhd";
